Enum libmodbus_rs::Exception

#[repr(C)]

pub enum Exception {
    ILLEGAL_FUNCTION,
    ILLEGAL_DATA_ADDRESS,
    ILLEGAL_DATA_VALUE,
    SLAVE_OR_SERVER_FAILURE,
    ACKNOWLEDGE,
    SLAVE_OR_SERVER_BUSY,
    NEGATIVE_ACKNOWLEDGE,
    MEMORY_PARITY,
    NOT_DEFINED,
    GATEWAY_PATH,
    GATEWAY_TARGET,
}

Modbus protocol exceptions

Documentation source: https://en.wikipedia.org/wiki/Modbus#Main_Modbus_exception_codes

Variants
ILLEGAL_FUNCTION

(1) Illegal Function - Function code received in the query is not recognized or allowed by slave

ILLEGAL_DATA_ADDRESS

(2) Illegal Data Address - Data address of some or all the required entities are not allowed or do not exist in slave

ILLEGAL_DATA_VALUE

(3) Illegal Data Value - Value is not accepted by slave

SLAVE_OR_SERVER_FAILURE

(4) Slave Device Failure - Unrecoverable error occurred while slave was attempting to perform requested action

ACKNOWLEDGE

(5) Acknowledge - Slave has accepted request and is processing it, but a long duration of time is required. This response is returned to prevent a timeout error from occurring in the master. Master can next issue a Poll Program Complete message to determine whether processing is completed

SLAVE_OR_SERVER_BUSY

(6) Slave Device Busy - Slave is engaged in processing a long-duration command. Master should retry later

NEGATIVE_ACKNOWLEDGE

(7) Negative Acknowledge - Slave cannot perform the programming functions. Master should request diagnostic or error information from slave

MEMORY_PARITY

(8) Memory Parity Error - Slave detected a parity error in memory. Master can retry the request, but service may be required on the slave device

NOT_DEFINED

(9) Not defined

GATEWAY_PATH

(10) Gateway Path Unavailable - Specialized for Modbus gateways. Indicates a misconfigured gateway

GATEWAY_TARGET

(11) Gateway Target Device Failed to Respond - Specialized for Modbus gateways. Sent when slave fails to respond