Struct libmodbus_rs::modbus::Modbus

pub struct Modbus { /* fields omitted */ }

This struct holds the current context ctx

This logic is derived from that one libmodbus uses.

Methods

impl Modbus

fn new_rtu(device: &str,
baud: i32,
parity: char,
data_bit: i32,
stop_bit: i32)
-> Self

Creates a new modbus context with the RTU backend

Attributes

device - Device string e.g. "/dev/ttyUSB0"

baud - Baud rate

parity - Parity

data_bit - Number of data bits

stop_bit - Number of stop bits

Examples

use libmodbus_rs::modbus::Modbus;

let modbus = Modbus::new_rtu("/dev/ttyUSB0", 9600, 'N', 8, 1);

fn set_slave(&mut self, slave_id: i32) -> Result<i32>

Define the slave ID of the remote device to talk in master mode or set the internal slave ID in slave mode

Attributes

slave_id - New modbus slave id (valid range: >= 0 && <= 247), 0 is broadcast address

Examples

A Ok(0) signals all right, on error a Error::InvalidSlaveID is returned

use libmodbus_rs::modbus::{Modbus, Error};

let mut modbus = Modbus::new_rtu("/dev/ttyUSB0", 9600, 'N', 8, 1);
assert_eq!(modbus.set_slave(10), Ok(0));
assert_eq!(modbus.set_slave(255), Err(Error::InvalidSlaveID));

fn set_debug(&mut self, flag: bool) -> Result<i32>

Set debug flag of the context

Attributes

flag - boolean true or false

Examples

A Ok(0) signals all right, on error a Error::GenericError is returned

use libmodbus_rs::modbus::{Modbus};

let mut modbus = Modbus::new_rtu("/dev/ttyUSB0", 9600, 'N', 8, 1);
assert_eq!(modbus.set_debug(true), Ok(0));

fn rtu_set_serial_mode(&mut self, mode: i32) -> Result<i32>

Set the serial mode

Attributes

mode - serial mode

Examples

A Ok(0) signals all right, on error a Error::GenericError is returned

use libmodbus_rs::modbus::{Modbus};

let mut modbus = Modbus::new_rtu("/dev/ttyUSB0", 9600, 'N', 8, 1);

fn rtu_set_rts(&mut self, mode: i32) -> Result<i32>

Set the RTS mode in RTU

Attributes

mode - serial mode

Examples

A Ok(0) signals all right, on error a Error::GenericError is returned

use libmodbus_rs::modbus::{Modbus};

let mut modbus = Modbus::new_rtu("/dev/ttyUSB0", 9600, 'N', 8, 1);
modbus.set_slave(1);
assert_eq!(modbus.rtu_set_rts(libmodbus_rs::MODBUS_RTU_RTS_DOWN), Ok(0));

fn connect(&self) -> Result<i32>

Establish a Modbus connection

Examples

A Ok(0) signals all right, on error a Error::GenericError is returned

use libmodbus_rs::modbus::{Modbus};

let mut modbus = Modbus::new_rtu("/dev/ttyUSB0", 9600, 'N', 8, 1);
modbus.set_slave(1);
assert_eq!(modbus.connect(), Ok(0));

fn close(&self)

modbus_close - close a Modbus connection

The modbus_close() function shall close the connection established with the backend set in the context.

Examples

use libmodbus_rs::modbus::{Modbus};

let mut modbus = Modbus::new_rtu("/dev/ttyUSB0", 9600, 'N', 8, 1);
modbus.set_slave(1);
match modbus.connect() {
    Err(_) => { modbus.close(); }
    Ok(_) => {
        let _ = modbus.read_registers(0, 1);
    }
}

fn free(&self)

Free a libmodbus context

Examples

use libmodbus_rs::modbus::{Modbus};

let mut modbus = Modbus::new_rtu("/dev/ttyUSB0", 9600, 'N', 8, 1);
modbus.set_slave(46);
let _ = modbus.connect();
modbus.free();

fn read_registers(&self, address: i32, num_reg: i32) -> Result<Vec<u16>>

Read many registers

Attributes

address - Start address from which the read should start

num_reg - Number of holding registers to read from address of the remote device

Examples

use libmodbus_rs::modbus::{Modbus};

let mut modbus = Modbus::new_rtu("/dev/ttyUSB0", 9600, 'N', 8, 1);
let _ = modbus.set_slave(46);
let _ = modbus.rtu_set_rts(libmodbus_rs::MODBUS_RTU_RTS_DOWN);
let _ = modbus.connect();
let mut tab_reg: Vec<u16> = modbus.read_registers(0, 19).unwrap();

fn read_input_registers(&self, address: i32, num_reg: i32) -> Result<Vec<u16>>

Read many registers

Attributes

address - Start address from which the read should start

num_reg - Number of holding registers to read from address of the remote device

Examples

use libmodbus_rs::modbus::{Modbus};

let mut modbus = Modbus::new_rtu("/dev/ttyUSB0", 9600, 'N', 8, 1);
let _ = modbus.set_slave(46);
let _ = modbus.rtu_set_rts(libmodbus_rs::MODBUS_RTU_RTS_DOWN);
let _ = modbus.connect();
let mut tab_reg: Vec<u16> = modbus.read_input_registers(0, 19).unwrap();

fn read_bits(&self, address: i32, num_bits: i32) -> Result<Vec<u8>>

read many bits

This function shall read the state of num_bits (coils) beginning from the address of the remote device. The return value is an vector of u8's encapsuled in a Result.

This funkciont uses the Modbus function 01 (0x01) Read Coils

Attributes

address - Start Adresse from where the read should begin

num_bits - Number of bit to be read

Examples

use libmodbus_rs::modbus::{Modbus};

let mut modbus = Modbus::new_rtu("/dev/ttyUSB0", 9600, 'N', 8, 1);
let _ = modbus.set_slave(46);
let _ = modbus.connect();
let _ = modbus.read_bits(0, 1).unwrap();

fn write_bit(&self, address: i32, state: bool) -> Result<()>

write a single coil

Attributes

address - Coil address

state - Coil state

Examples

use libmodbus_rs::modbus::{Modbus};

let mut modbus = Modbus::new_rtu("/dev/ttyUSB0", 9600, 'N', 8, 1);
let _ = modbus.set_slave(46);
let _ = modbus.connect();
let _ = modbus.write_bit(0, true).unwrap();

fn write_register(&self, address: i32, value: i32) -> Result<()>

write a single register

Attributes

address - Register address

value - New register value

Examples

use libmodbus_rs::modbus::{Modbus};

let mut modbus = Modbus::new_rtu("/dev/ttyUSB0", 9600, 'N', 8, 1);
let _ = modbus.set_slave(46);
let _ = modbus.connect();
let _ = modbus.write_register(0, 0x1234);
let tab_reg: Vec<u16> = modbus.read_registers(0, 1).unwrap();
assert_eq!(tab_reg, &[0x1234]);

fn write_registers(&self, address: i32, register: &[u16]) -> Result<i32>

write multiple registers

Attributes

address - Start address register

register - Vector of u16 values

Examples

use libmodbus_rs::modbus::{Modbus};

let mut modbus = Modbus::new_rtu("/dev/ttyUSB0", 9600, 'N', 8, 1);
let _ = modbus.set_slave(46);
let _ = modbus.connect();
let register = [0x1111u16, 0x2222u16];
assert_eq!(modbus.write_registers(0, &register), Ok(2));

fn send_raw_request(&self, raw_request: &[u8]) -> Result<i32>

send a raw request

Attributes

raw_request - Vector of u8 values

Examples

use libmodbus_rs::modbus::{Modbus};

let mut modbus = Modbus::new_rtu("/dev/ttyUSB0", 9600, 'N', 8, 1);
let _ = modbus.set_slave(46);
let _ = modbus.connect();
let raw_request =  vec![0x2E, 0x08, 0x00, 0x01];
assert_eq!(modbus.send_raw_request(&raw_request), Ok(6));