Struct libmodbus_rs::Modbus

pub struct Modbus {
    pub ctx: *mut modbus_t,
}

Safe interface for libmodbus

The different parts of libmodbus are implemented as traits. The modules of this crate contains these traits and a implementation with a, hopefully safe, interface.

Fields
ctx: *mut modbus_t
Methods

impl Modbus

const ENOBASE: u32

ENOBASE: u32 = ffi::MODBUS_ENOBASE

const MAX_ADU_LENGTH: usize

MAX_ADU_LENGTH: usize = ffi::MODBUS_MAX_ADU_LENGTH as usize

const MAX_PDU_LENGTH: usize

MAX_PDU_LENGTH: usize = ffi::MODBUS_MAX_PDU_LENGTH as usize

const MAX_READ_BITS: usize

MAX_READ_BITS: usize = ffi::MODBUS_MAX_READ_BITS as usize

const MAX_READ_REGISTERS: usize

MAX_READ_REGISTERS: usize = ffi::MODBUS_MAX_READ_REGISTERS as usize

const MAX_WR_READ_REGISTERS: usize

MAX_WR_READ_REGISTERS: usize = ffi::MODBUS_MAX_WR_READ_REGISTERS as usize

const MAX_WR_WRITE_REGISTERS: usize

MAX_WR_WRITE_REGISTERS: usize = ffi::MODBUS_MAX_WR_WRITE_REGISTERS as usize

const MAX_WRITE_BITS: usize

MAX_WRITE_BITS: usize = ffi::MODBUS_MAX_WRITE_BITS as usize

const MAX_WRITE_REGISTERS: usize

MAX_WRITE_REGISTERS: usize = ffi::MODBUS_MAX_WRITE_REGISTERS as usize

const RTU_MAX_ADU_LENGTH: usize

RTU_MAX_ADU_LENGTH: usize = ffi::MODBUS_RTU_MAX_ADU_LENGTH as usize

const TCP_DEFAULT_PORT: u32

TCP_DEFAULT_PORT: u32 = ffi::MODBUS_TCP_DEFAULT_PORT

const TCP_MAX_ADU_LENGTH: u32

TCP_MAX_ADU_LENGTH: u32 = ffi::MODBUS_TCP_MAX_ADU_LENGTH

const TCP_SLAVE: u32

TCP_SLAVE: u32 = ffi::MODBUS_TCP_SLAVE

pub fn connect(&self) -> Result<()>

connect - establish a Modbus connection

The connect() function shall establish a connection to a Modbus server, a network or a bus.

Return value

The function return an OK Result if successful. Otherwise it contains an Error.

Examples

use libmodbus_rs::{Modbus, ModbusTCP};

// create server
let mut server = Modbus::new_tcp("127.0.0.1", 1502).unwrap();
// create client
let client = Modbus::new_tcp("127.0.0.1", 1502).unwrap();
// start server in listen mode
let _ = server.tcp_listen(1).unwrap();

assert!(client.connect().is_ok())

pub fn flush(&self) -> Result<()>

flush - flush non-transmitted data

The flush() function shall discard data received but not read to the socket or file descriptor associated to the context ctx.

Return value

The function return an OK Result if successful. Otherwise it contains an Error.

Examples

use libmodbus_rs::{Modbus, ModbusTCP};
let modbus = Modbus::new_tcp("127.0.0.1", 1502).unwrap();

assert!(modbus.flush().is_ok());

pub fn set_slave(&mut self, slave: u8) -> Result<()>

set_slave - set slave number in the context

The set_slave() function shall set the slave number in the libmodbus context. The behavior depends of network and the role of the device:

RTU Define the slave ID of the remote device to talk in master mode or set the internal slave ID in slave mode. According to the protocol, a Modbus device must only accept message holding its slave number or the special broadcast number. TCP The slave number is only required in TCP if the message must reach a device on a serial network. Some not compliant devices or software (such as modpoll) uses the slave ID as unit identifier, that’s incorrect (cf page 23 of Modbus Messaging Implementation Guide v1.0b) but without the slave value, the faulty remote device or software drops the requests! The special value MODBUS_TCP_SLAVE (0xFF) can be used in TCP mode to restore the default value. The broadcast address is MODBUS_BROADCAST_ADDRESS. This special value must be use when you want all Modbus devices of the network receive the request.

Return value

The function return an OK Result if successful. Otherwise it contains an Error.

Parameters

slave - new slave ID

Examples

use libmodbus_rs::{Modbus, ModbusRTU};

const YOUR_DEVICE_ID: u8 = 1;
let mut modbus = Modbus::new_rtu("/dev/ttyUSB0", 115200, 'N', 8, 1).unwrap();

assert!(modbus.set_slave(YOUR_DEVICE_ID).is_ok());

pub fn set_debug(&mut self, flag: bool) -> Result<()>

set_debug - set debug flag of the context

The set_debug() function shall set the debug flag of the modbus_t context by using the argument flag. By default, the boolean flag is set to FALSE. When the flag value is set to TRUE, many verbose messages are displayed on stdout and stderr. For example, this flag is useful to display the bytes of the Modbus messages.

[00][14][00][00][00][06][12][03][00][6B][00][03]
Waiting for a confirmation…
<00><14><00><00><00><09><12><03><06><02><2B><00><00><00><00>

Return value

The function return an OK Result if successful. Otherwise it contains an Error.

Parameters

flag - true of false, enables or disables debug mode

Examples

use libmodbus_rs::{Modbus, ModbusTCP};

let mut modbus = Modbus::new_tcp("127.0.0.1", 1502).unwrap();

assert!(modbus.set_debug(true).is_ok());

pub fn get_byte_timeout(&self) -> Result<Timeout>

get_byte_timeout - get timeout between bytes

get_byte_timeout() function returns a Timeout with the timeout interval between two consecutive bytes of the same message.

Return value

The function return a Result containing a Timeout if successful. Otherwise it contains an Error.

Examples

use libmodbus_rs::{Modbus, ModbusTCP, Timeout};
let mut modbus = Modbus::new_tcp("127.0.0.1", 1502).unwrap();

assert_eq!(modbus.get_byte_timeout().unwrap(), Timeout { sec: 0, usec: 500000 });

pub fn set_byte_timeout(&mut self, timeout: Timeout) -> Result<()>

set_byte_timeout - set timeout between bytes

The set_byte_timeout() function shall set the timeout interval between two consecutive bytes of the same message. The timeout is an upper bound on the amount of time elapsed before select() returns, if the time elapsed is longer than the defined timeout, an ETIMEDOUT error will be raised by the function waiting for a response.

The value of usec argument must be in the range 0 to 999999.

If both sec and usec are zero, this timeout will not be used at all. In this case, set_byte_timeout() governs the entire handling of the response, the full confirmation response must be received before expiration of the response timeout. When a byte timeout is set, the response timeout is only used to wait for until the first byte of the response.

Return value

The function return an OK Result if successful. Otherwise it contains an Error.

Parameters

timeout - Timeout struct with sec and usec

Examples

use libmodbus_rs::{Modbus, ModbusTCP, Timeout};
let mut modbus = Modbus::new_tcp("127.0.0.1", 1502).unwrap();
let timeout = Timeout { sec: 1, usec: 500000 };
assert!(modbus.set_byte_timeout(timeout).is_ok());

pub fn get_response_timeout(&self) -> Result<Timeout>

get_response_timeout - get timeout for response

The get_response_timeout() function shall return the timeout interval used to wait for a response in the sec and usec arguments.

Return value

The function return a Result containing a Timeout if successful. Otherwise it contains an Error.

Examples

use libmodbus_rs::{Modbus, ModbusTCP, Timeout};
let mut modbus = Modbus::new_tcp("127.0.0.1", 1502).unwrap();

assert_eq!(modbus.get_response_timeout().unwrap(), Timeout { sec: 0, usec: 500000 });

pub fn set_response_timeout(&mut self, timeout: Timeout) -> Result<()>

set_response_timeout - set timeout for response

The set_response_timeout() function shall set the timeout interval used to wait for a response. When a byte timeout is set, if elapsed time for the first byte of response is longer than the given timeout, an ETIMEDOUT error will be raised by the function waiting for a response. When byte timeout is disabled, the full confirmation response must be received before expiration of the response timeout.

If the Timeout members are both sec and usec are zero, this timeout will not be used at all. In this case, set_response_timeout() governs the entire handling of the response, the full confirmation response must be received before expiration of the response timeout. When a byte timeout is set, the response timeout is only used to wait for until the first byte of the response.

Return value

The function return an OK Result if successful. Otherwise it contains an Error.

Parameters

Timeout - Timeout

Examples

use libmodbus_rs::{Modbus, ModbusTCP, Timeout};
let mut modbus = Modbus::new_tcp("127.0.0.1", 1502).unwrap();
let timeout = Timeout { sec: 1, usec: 500000 };
assert!(modbus.set_response_timeout(timeout).is_ok());

pub fn set_error_recovery(
&mut self,
flags: Option<&[ErrorRecoveryMode]>
) -> Result<()>

set_error_recovery - set the error recovery mode

The set_error_recovery() function shall set the error recovery mode to apply when the connection fails or the byte received is not expected.

By default there is no error recovery so the application is responsible for controlling the error values returned by libmodbus functions and for handling them if necessary.

When ErrorRecoveryMode::Link is set, the library will attempt an reconnection after a delay defined by response timeout (set_response_timeout()) of the libmodbus context. This mode will try an infinite close/connect loop until success on send call and will just try one time to re-establish the connection on select/read calls (if the connection was down, the values to read are certainly not available any more after reconnection, except for slave/server). This mode will also run flush requests after a delay based on the current response timeout in some situations (eg. timeout of select call). The reconnection attempt can hang for several seconds if the network to the remote target unit is down.

When ErrorRecoveryMode::PROTOCOL is set, a sleep and flush sequence will be used to clean up the ongoing communication, this can occurs when the message length is invalid, the TID is wrong or the received function code is not the expected one. The response timeout delay will be used to sleep.

The modes are mask values and so they are complementary.

It’s not recommended to enable error recovery for slave/server.

Return value

The function return an OK Result if successful. Otherwise it contains an Error.

Parameters

ErrorRecoveryMode - Timeout

Examples

use libmodbus_rs::{Modbus, ModbusTCP, ErrorRecoveryMode};
let mut modbus = Modbus::new_tcp("127.0.0.1", 1502).unwrap();

assert!(modbus.set_error_recovery(Some(&[ErrorRecoveryMode::Link, ErrorRecoveryMode::Protocol])).is_ok());

pub fn set_socket(&mut self, socket: i32) -> Result<()>

set_socket - set socket of the context

The set_socket() function shall set the socket or file descriptor in the libmodbus context. This function is useful for managing multiple client connections to the same server.

Return values

The function return an OK Result if successful. Otherwise it contains an Error.

Examples

use libmodbus_rs::{Modbus, ModbusTCP};
let mut modbus = Modbus::new_tcp("127.0.0.1", 1502).unwrap();

assert!(modbus.set_socket(1337).is_ok());

pub fn get_socket(&self) -> Result<i32>

get_socket - set socket of the context

The get_socket() function shall return the current socket or file descriptor of the libmodbus context.

Return value

The function returns a Result containing the current socket or file descriptor of the context if successful. Otherwise it contains an Error.

Examples

use libmodbus_rs::{Modbus, ModbusTCP};
let mut modbus = Modbus::new_tcp("127.0.0.1", 1502).unwrap();
let _ = modbus.set_socket(1337).unwrap();
assert_eq!(modbus.get_socket().unwrap(), 1337);

pub fn get_header_length(&self) -> i32

get_header_length - retrieve the current header length

The get_header_length() function shall retrieve the current header length from the backend. This function is convenient to manipulate a message and so its limited to low-level operations.

Return values

The header length as integer value.

Examples

use libmodbus_rs::{Modbus, ModbusTCP};
let modbus = Modbus::new_tcp("127.0.0.1", 1502).unwrap();
assert_eq!(modbus.get_header_length(), 7);

pub fn reply_exception(
&self,
request: &[u8],
exception_code: u32
) -> Result<i32>

reply_exception - send an exception reponse

The modbus_reply_exception() function shall send an exception response based on the exception_code in argument.

The libmodbus provides the following exception codes:

Modbus::Exception::IllegalFunction (1)

Modbus::Exception::IllegalDataAddress (2)

Modbus::Exception::IllegalDataValue (3)

Modbus::Exception::SlaveOrServerFailure (4)

Modbus::Exception::Acknowledge (5)

Modbus::Exception::SlaveDeviceBusy (6)

Modbus::Exception::NegativeAcknowledge (7)

Modbus::Exception::MemoryParity (8)

Modbus::Exception::NotDefined (9)

Modbus::Exception::GatewayPath (10)

Modbus::Exception::GatewayTarget (11)

The initial request request is required to build a valid response.

Return value

The function returns the length of the response sent if successful, or an Error.

Parameters

request - initial request, required to build a valid response

exception_code - Exception Code

Examples

use libmodbus_rs::{Modbus, ModbusClient, ModbusTCP};
let modbus = Modbus::new_tcp("127.0.0.1", 1502).unwrap();
use libmodbus_rs::Exception;

let request: Vec<u8> = vec![0x01];
assert_eq!(modbus.reply_exception(&request, Exception::Acknowledge as u32).unwrap(), 9);

pub fn close(&self)

close - close a Modbus connection

The close() function shall close the connection established with the backend set in the context.

It should not nessesary to call these function. Because rusts drop trait handles that for you!

Examples

use libmodbus_rs::{Modbus, ModbusTCP};
let mut modbus = Modbus::new_tcp("127.0.0.1", 1502).unwrap();
modbus.close();

pub fn free(&mut self)

free - free a libmodbus context

The free() function shall free an allocated modbus_t structure.

It should not nessesary to call these function. Because rusts drop trait handles that for you!

Examples

use libmodbus_rs::{Modbus, ModbusTCP};
let mut modbus = Modbus::new_tcp("127.0.0.1", 1502).unwrap();
modbus.free();