Struct libmodbus_rs::ModbusMapping

pub struct ModbusMapping {
    pub modbus_mapping: *mut modbus_mapping_t,
}

To handle the mapping of your Modbus data, you must use this struct

Fields
modbus_mapping: *mut modbus_mapping_t
Methods

impl ModbusMapping

fn new(number_bits: i32,
number_input_bits: i32,
number_registers: i32,
number_input_registers: i32)
-> Result<ModbusMapping>

new - allocate four arrays of bits and registers

Parameters

number_bits - How many bits sould allocated

number_input_bits - How many bits sould allocated

number_registers - How many registers sould allocated

number_input_registers - How many input registers sould allocated

Examples

use libmodbus_rs::{Modbus, ModbusMapping, ModbusTCP};
let modbus = Modbus::new_tcp("127.0.0.1", 1502).unwrap();

let modbus_mapping = ModbusMapping::new(500, 500, 500, 500).unwrap();

fn free(&mut self)

mapping_free - free a modbus_mapping_t structure

The function shall free the four arrays of mb_mapping_t structure and finally the mb_mapping_t referenced by mb_mapping.

It should not nessesary to call these function. Because rusts drop trait handles that for you!

Examples

use libmodbus_rs::ModbusMapping;

let mut modbus_mapping = ModbusMapping::new(500, 500, 500, 500).unwrap();

modbus_mapping.free();